+++ packages/drafter/src/refract/JsonSchema.cc
@@ -428,16 +428,25 @@ namespace
 
         if (options.test(FIXED_TYPE_FLAG)) { // array of any of types
 
-            so::Array items{};
-            if (!e.empty())
+            auto& schema = wrapNullable(s, options);
+            addType(schema, TYPE_NAME);
+
+            if (e.empty()) {
+                so::Array items{};
+                addItems(schema, so::Object{ so::from_list{}, std::make_pair("anyOf", std::move(items)) });
+            } else if (e.get().size() == 1) {
+                const auto& entry = *e.get().begin();
+                so::Object items = makeSchema(*entry, inheritOrPassFlags(options, *entry));
+                addItems(schema, std::move(items));
+            } else {
+                so::Array items{};
                 for (const auto& entry : e.get()) {
                     assert(entry);
                     so::emplace_unique(items, makeSchema(*entry, inheritOrPassFlags(options, *entry)));
                 }
 
-            auto& schema = wrapNullable(s, options);
-            addType(schema, TYPE_NAME);
-            addItems(schema, so::Object{ so::from_list{}, std::make_pair("anyOf", std::move(items)) });
+                addItems(schema, so::Object{ so::from_list{}, std::make_pair("anyOf", std::move(items)) });
+            }
 
         } else if (options.test(FIXED_FLAG)) { // tuple of N constants/types