+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Drafter Changelog
 
-## TBD
+## 5.1.0 (2023-05-17)
 
 ### Enhancements